+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_component_showcase.tsx
@@ -17,9 +17,9 @@ import { CWRadioButton } from './cw_radio_button';
 import { CWWalletOptionRow } from './cw_wallet_option_row';
 import { CWAccountCreationButton } from './cw_account_creation_button';
 import { CWCheckbox } from './cw_checkbox';
-import { CWTooltip } from './cw_tooltip';
+import { CWTooltip } from './cw_popover/cw_tooltip';
 import { CWPopover } from './cw_popover/cw_popover';
-import { CWAddressTooltip } from './cw_address_tooltip';
+import { CWAddressTooltip } from './cw_popover/cw_address_tooltip';
 import { ValidationStatus } from './cw_validation_text';
 import { CWTextArea } from './cw_text_area';
 import { CWTab, CWTabBar } from './cw_tabs';
@@ -78,6 +78,113 @@ export class ComponentShowcase implements m.ClassComponent {
             })
           }
         /> */}
+        <h1>Popover</h1>
+        <div class="basic-gallery">
+          <CWPopover
+            trigger={<CWIconButton iconName="dots" />}
+            content={<>Some content</>}
+            interactionType="click"
+          />
+        </div>
+        <h1>Tooltip</h1>
+        <div class="tooltip-gallery">
+          <div class="tooltip-row">
+            <CWText>Hover</CWText>
+            <CWTooltip
+              interactionType="hover"
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="bordered"
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Hover to side</CWText>
+            <CWTooltip
+              interactionType="hover"
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="bordered"
+              toSide
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Persist on hover</CWText>
+            <CWTooltip
+              trigger={<CWIcon iconName="infoEmpty" />}
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="bordered"
+              interactionType="hover"
+              persistOnHover
+              hoverCloseDelay={1500}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Hover open and close delay</CWText>
+            <CWTooltip
+              trigger={<CWIcon iconName="infoEmpty" />}
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="bordered"
+              interactionType="hover"
+              hoverOpenDelay={1500}
+              hoverCloseDelay={1500}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Click</CWText>
+            <CWTooltip
+              interactionType="click"
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="bordered"
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Solid background</CWText>
+            <CWTooltip
+              interactionType="hover"
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="solidArrow"
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Solid background, no arrow</CWText>
+            <CWTooltip
+              interactionType="hover"
+              tooltipContents={`
+                I am an informational tool tip here to provide \
+                extra details on things people may need more help on.
+              `}
+              tooltipType="solidNoArrow"
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+          <div class="tooltip-row">
+            <CWText>Address tooltip</CWText>
+            <CWAddressTooltip
+              address="0xa5430730f12f1128bf10dfba38c8e00bc4d90eea"
+              trigger={<CWIcon iconName="infoEmpty" />}
+            />
+          </div>
+        </div>
         <div class="basic-gallery">
           <h1>Toggle</h1>
           <CWToggle
@@ -189,113 +296,6 @@ export class ComponentShowcase implements m.ClassComponent {
             walletName="metamask"
             onclick={() => notifySuccess('MetaMask clicked!')}
           />
-        </div>
-        <h1>Popover</h1>
-        <div class="tooltip-gallery">
-          <CWPopover
-            trigger={<CWButton label="Click me" />}
-            content={<>Some content</>}
-            interactionType="click"
-          />
-        </div>
-        <h1>Tooltip</h1>
-        <div class="tooltip-gallery">
-          <div class="tooltip-row">
-            <CWText>Hover</CWText>
-            <CWTooltip
-              interactionType="hover"
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="bordered"
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Hover to side</CWText>
-            <CWTooltip
-              interactionType="hover"
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="bordered"
-              toSide
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Persist on hover</CWText>
-            <CWTooltip
-              trigger={<CWIcon iconName="infoEmpty" />}
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="bordered"
-              interactionType="hover"
-              persistOnHover
-              hoverCloseDelay={1500}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Hover open and close delay</CWText>
-            <CWTooltip
-              trigger={<CWIcon iconName="infoEmpty" />}
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="bordered"
-              interactionType="hover"
-              hoverOpenDelay={1500}
-              hoverCloseDelay={1500}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Click</CWText>
-            <CWTooltip
-              interactionType="click"
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="bordered"
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Solid background</CWText>
-            <CWTooltip
-              interactionType="hover"
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="solidArrow"
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Solid background, no arrow</CWText>
-            <CWTooltip
-              interactionType="hover"
-              tooltipContents={`
-                I am an informational tool tip here to provide \
-                extra details on things people may need more help on.
-              `}
-              tooltipType="solidNoArrow"
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
-          <div class="tooltip-row">
-            <CWText>Address tooltip</CWText>
-            <CWAddressTooltip
-              address="0xa5430730f12f1128bf10dfba38c8e00bc4d90eea"
-              trigger={<CWIcon iconName="infoEmpty" />}
-            />
-          </div>
         </div>
         <h1>Form fields</h1>
         <div class="form-gallery">
