+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_popover/cw_address_tooltip.tsx
@@ -2,13 +2,13 @@
 
 import m from 'mithril';
 
-import 'components/component_kit/cw_address_tooltip.scss';
+import 'components/component_kit/cw_popover/cw_address_tooltip.scss';
 
 import { notifyError, notifySuccess } from 'controllers/app/notifications';
+import { CWIconButton } from '../cw_icon_button';
+import { CWText } from '../cw_text';
+import { ComponentType } from '../types';
 import { CWTooltip } from './cw_tooltip';
-import { ComponentType } from './types';
-import { CWText } from './cw_text';
-import { CWIconButton } from './cw_icon_button';
 
 export class CWAddressTooltip implements m.ClassComponent<{ address: string }> {
   view(vnode) {
