+++ packages/commonwealth/client/styles/mixins/popovers.scss
@@ -4,6 +4,18 @@
   z-index: 1001;
 }
 
+@mixin baseTooltipStyles {
+  align-items: center;
+  background-color: $white;
+  border-radius: $border-radius-corners;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  overflow: hidden;
+  max-width: 196px;
+  min-height: 8px;
+}
+
 @mixin baseArrowStyles {
   height: 0;
   position: absolute;
